backend/app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.7.2
motor==3.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.7.0
